+++ cg/cli/clean.py
@@ -28,7 +28,7 @@ from cg.constants import FlowCellStatus
 from cg.constants.constants import DRY_RUN, SKIP_CONFIRMATION
 from cg.constants.sequencing import Sequencers
 from cg.constants.housekeeper_tags import SequencingFileTag, ALIGNMENT_FILE_TAGS, ScoutTag
-from cg.datetime.utils import get_date_days_ago, get_timedelta_from_date
+from cg.utils.date import get_timedelta_from_date, get_date_days_ago
 from cg.exc import FlowCellError
 from cg.meta.clean.api import CleanAPI
 from cg.meta.clean.demultiplexed_flow_cells import DemultiplexedRunsFlowCell
