+++ cg/constants/__init__.py
@@ -25,5 +25,5 @@ from cg.constants.priority import Priority
 from cg.constants.process import EXIT_FAIL, EXIT_SUCCESS, RETURN_SUCCESS
 from cg.constants.report import *
 from cg.constants.sample_sources import ANALYSIS_SOURCES, METAGENOME_SOURCES
-from cg.constants.symbols import NO_PARENT, SINGLE_QUOTE, SPACE
+from cg.constants.symbols import SINGLE_QUOTE, SPACE
 from cg.constants.housekeeper_tags import HK_FASTQ_TAGS
