+++ cg/constants/symbols.py
@@ -1,11 +1,12 @@
-"""Symbol constants"""
+"""Symbol constants."""
 
+ASTERISK = "*"
+COMMA = ","
+DASH = "-"
+DOT = "."
+FWD_SLASH = "/"
+NEW_LINE = "\n"
+PERIOD = "."
 SINGLE_QUOTE = "'"
 SPACE = " "
 UNDERSCORE = "_"
-PERIOD = "."
-COMMA = ","
-ASTERISK = "*"
-NEW_LINE = "\n"
-
-NO_PARENT = "0"
